MyanmarNLPTools/MyanmarNLPTools/requirements.txt
myanmartools
PyICU
